refactor: split out gradio ui
